requirements.txt
google-cloud-bigquery==3.13.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-cloud-storage==2.10.0
pandas==1.5.3
python-dotenv==1.0.0
